Update WHMCS order status when customer redirected back to success_redirect_url
